requirements.txt
Django>=5.1
djangorestframework
psycopg2
django-cors-headers
